+++ inertia/pages/settings/billing.vue
@@ -6,10 +6,16 @@ export default { layout: DefaultLayout }
 <script setup lang="ts">
 import SettingsShell from '~/components/settings/SettingsShell.vue'
 import SettingsBillingTab from '~/components/settings/tabs/SettingsBillingTab.vue'
+import type { PlanTier, QuotaUsage } from '../../../shared/types/plan'
+
+defineProps<{
+  plan: PlanTier
+  quotaUsage: QuotaUsage
+}>()
 </script>
 
 <template>
   <SettingsShell>
-    <SettingsBillingTab />
+    <SettingsBillingTab :plan="plan" :quota-usage="quotaUsage" />
   </SettingsShell>
 </template>